+++ components/index.js
@@ -37,6 +37,8 @@ import CellItem from './cell-item'
 import Switch from './switch'
 import Agree from './agree'
 import Radio from './radio'
+import RadioGroup from './radio-group'
+import RadioBox from './radio-box'
 import RadioList from './radio-list'
 import DatePicker from './date-picker'
 import Captcha from './captcha'
@@ -59,6 +61,8 @@ import DetailItem from './detail-item'
 import Slider from './slider'
 import Progress from './progress'
 import Ruler from './ruler'
+import TextareaItem from './textarea-item'
+import Skeleton from './skeleton'
 /* @init<%import ${componentNameUpper} from './${componentName}'%> */
 
 // Totally importing reminder
@@ -104,6 +108,8 @@ export const components = {
   Switch,
   Agree,
   Radio,
+  RadioGroup,
+  RadioBox,
   RadioList,
   DatePicker,
   Captcha,
@@ -127,6 +133,8 @@ export const components = {
   Slider,
   Progress,
   Ruler,
+  TextareaItem,
+  Skeleton,
   /* @init<%${componentNameUpper},%> */
 }
 
@@ -193,6 +201,8 @@ export {
   Switch,
   Agree,
   Radio,
+  RadioGroup,
+  RadioBox,
   RadioList,
   DatePicker,
   Captcha,
@@ -216,6 +226,8 @@ export {
   Slider,
   Progress,
   Ruler,
+  TextareaItem,
+  Skeleton,
   /* @init<%${componentNameUpper},%> */
 }