+++ components/skeleton/test/cases/demo3.vue
@@ -1,0 +1,22 @@
+<template>
+  <div class="md-example">
+    <md-skeleton :loading="false">
+      <h4></h4>
+    </md-skeleton>
+  </div>
+</template>
+
+<script>import {Skeleton} from 'mand-mobile'
+export default {
+  name: 'skeleton-demo',
+  title: '组合骨架',
+  data() {
+    return {
+      loading: true,
+    }
+  },
+  components: {
+    [Skeleton.name]: Skeleton,
+  },
+}
+</script>