+++ components/textarea-item/component.js
@@ -1,0 +1,7 @@
+export default {
+  'name': 'textarea-item',
+  'text': '文本输入框',
+  'category': 'form',
+  'description': '文本输入框。',
+  'author': '张鹏'
+}